backend/requirements.txt
# Core & API
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2
fastapi-cli==0.0.7

# Data & Validation
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.9.1
annotated-types==0.7.0
typing-inspection==0.4.0
typing_extensions==4.13.2
# Database
pymongo==4.13.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
psycopg2-binary==2.9.10
# Utilities
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
python-decouple==3.8
python-jose==3.4.0
passlib==1.7.4
# HTTP & Networking
httpx==0.28.1
requests==2.32.3
dnspython==2.7.0
idna==3.10
certifi==2025.1.31
httpcore==1.0.9
httptools==0.6.4
websockets==15.0.1
# CLI & Output
click==8.1.8
typer==0.16.0
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
# Jinja & Markdown
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.19.1
# Async & Misc
anyio==4.9.0
sniffio==1.3.1
exceptiongroup==1.2.2
watchfiles==1.0.5
# IPython/Jupyter (opcional, para notebooks)
ipykernel==6.29.5
ipython==8.36.0
jupyter_client==8.6.3
jupyter_core==5.8.1
traitlets==5.14.3
matplotlib-inline==0.1.7
prompt_toolkit==3.0.51
stack-data==0.6.3
pure_eval==0.2.3
asttokens==3.0.0
parso==0.8.4
jedi==0.19.2
decorator==5.2.1
comm==0.2.2
nest-asyncio==1.6.0
pyzmq==26.4.0
# Otros útiles y multiplataforma
colorama==0.4.4
platformdirs==4.3.8
six==1.16.0
packaging==24.2
wcwidth==0.2.13
zipp==1.0.0
# --- Dependencias problemáticas en Windows, comentadas ---
# dbus-python==1.2.18
# systemd-python==234
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
# python-apt==2.4.0+ubuntu3
# distro-info==1.1+ubuntu0.2
# wadllib==1.3.6
# launchpadlib==1.10.16
# lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
# PyGObject==3.42.1
# gunicorn==23.0.0
# pexpect==4.9.0
# ptyprocess==0.7.0
# greenlet==3.2.0
# asgiref==3.5.0
## The following requirements were added by pip freeze:
asgiref==3.5.0
bcrypt==4.0.1
blinker==1.4
charset-normalizer==3.4.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.8.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
greenlet==3.2.0
gunicorn==23.0.0
gyp==0.1
h11==0.16.0
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
opengateway-sandbox-sdk==1.1.0
pexpect==4.9.0
psutil==7.0.0
ptyprocess==0.7.0
pyasn1==0.4.8
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
rsa==4.9.1
SecretStorage==3.3.1
systemd-python==234
tornado==6.5.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.4.0
uvloop==0.21.0
wadllib==1.3.6
wsproto==1.0.0
## The following requirements were added by pip freeze:
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.63
langchain-openai==0.3.18
langsmith==0.3.43
openai==1.82.1
orjson==3.10.18
regex==2024.11.6
requests-toolbelt==1.0.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
zstandard==0.23.0
numpy